- nullptr guard for dlSymsInit for pe and elf, for coherency with free(), etc. - rbdc cleanups - changelog entry about new rbdc
git-svn-id: http://dyncall.org/svn/dyncall/trunk@1556 c71cb566-5a31-4692-9165-cf76bdc42cab

+++ bindings/ruby/rbdc/examples/unix_math.rb
@@ -15,10 +15,11 @@ if l.load('/usr/lib/libm.so') != nil and l.syms_init('/usr/lib/libm.so') != nil
   puts 'All symbol names in libm:'
   l.syms_each { |s| puts s }
 
-  puts 'libm has sqrtf()? '      +l.exists?(:sqrtf)      .to_s
-  puts 'libm has pow()? '        +l.exists?(:pow)        .to_s
-  puts 'libm has not_in_libm()? '+l.exists?(:not_in_libm).to_s
-  puts 'sqrtf(36.f) = '+l.call(:sqrt, 'f)f', 36.0).to_s
-  puts 'pow(2.0, 10.0) = '+l.call(:pow, 'dd)d', 2.0, 10.0).to_s
+  puts 'libm has sqrtf()? '       + l.exists?(:sqrtf)      .to_s
+  puts 'libm has pow()? '         + l.exists?(:pow)        .to_s
+  puts 'libm has not_in_libm()? ' + l.exists?(:not_in_libm).to_s
+
+  puts 'sqrtf(36.f) = '    + l.call(:sqrtf, 'f)f', 36.0)      .to_s
+  puts 'pow(2.0, 10.0) = ' + l.call(:pow,   'dd)d', 2.0, 10.0).to_s
 end
